# api/swagger.yaml
openapi: 3.0.0
info:
  title: HerdMaster
  description: |-
    This is API provided by the HerdMaster application. No shepherd dog were missed.
    
    Some useful links:
    - [The Ram DB repository](https://github.com/KennyMacCormik/HerdMaster)
    - [The source API definition for the Ram DB](https://github.com/KennyMacCormik/HerdMaster/api/swagger.yaml)
  version: 1.0.0
tags:
  - name: dog
    description: implements basic CRUD operations for the shepherd_dogs
  - name: owner
    description: implements basic CRUD operations for the owners
paths:
  /owner/{id}:
    get:
      tags:
        - owner
      summary: Find owner by ID
      description: Returns an owner
      parameters:
        - name: id
          in: path
          description: ID of an owner to return
          required: true
          schema:
            type: integer
            format: int64
            example: 21
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - created_at
                  - updated_at
                  - name
                  - email
                  - phone
                  - coat_id
                  - address_id
                properties:
                  id:
                    type: integer
                    format: int64
                    description: "Primary key, unique identifier for each dog"
                    example: 102
                  created_at:
                    type: integer
                    format: int64
                    description: "Record creation timestamp"
                    example: 1729614468
                  updated_at:
                    type: integer
                    format: int64
                    description: "Last updated timestamp"
                    example: 1729614468
                  name:
                    type: string
                    description: "Owner’s full name"
                    example: "Alex Ivanov"
                  email:
                    type: string
                    format: email
                    description: "Email address"
                    example: "owners.email@gmail.com"
                  phone:
                    type: string
                    description: "Contact phone number"
                    example: "+71234567890"
                  address_id:
                    type: integer
                    format: int64
                    description: "Foreign key, references addresses table"
                    example: 61
        '400':
          $ref: "#/components/responses/InvalidRequest"
        '404':
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - owner
      summary: Creates or updates an owner by ID
      description: Creates or updates an owner
      parameters:
        - name: id
          in: path
          description: ID of an owner to create or update
          required: true
          schema:
            type: integer
            format: int64
            example: 14
      requestBody:
        required: true
        description: Creates or updates an owner by ID
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "Owner’s full name"
                  example: "Alex Ivanov"
                email:
                  type: string
                  description: "Email address"
                  example: "owners.email@gmail.com"
                phone:
                  type: string
                  description: "Contact phone number"
                  example: "+71234567890"
                address_id:
                  type: integer
                  format: int64
                  description: "Foreign key, references addresses table"
                  example: 61
      responses:
        '200':
          description: Successful operation
        '201':
          description: Successful creation
        '400':
          $ref: "#/components/responses/InvalidRequest"
    delete:
      tags:
        - owner
      summary: Delete an owner by ID
      description: Deletes an owner
      parameters:
        - name: id
          in: path
          description: ID of an owner to delete
          required: true
          schema:
            type: integer
            format: int64
            example: 14
      responses:
        '204':
          description: Successful operation
        '400':
          $ref: "#/components/responses/InvalidRequest"
        '404':
          $ref: "#/components/responses/NotFound"
  /dog/{id}:
    get:
      tags:
        - dog
      summary: Find shepherd dog by ID
      description: Returns a shepherd dog
      parameters:
        - name: id
          in: path
          description: ID of a shepherd dog to return
          required: true
          schema:
            type: integer
            format: int64
            example: 21
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - created_at
                  - updated_at
                  - owner_id
                  - breed_id
                  - gender_id
                  - coat_id
                  - name
                  - weight_g
                  - height_mm
                  - microchip_number
                  - vaccinated
                  - date_of_birth
                properties:
                  id:
                    type: integer
                    format: int64
                    description: "Primary key, unique identifier for each dog"
                    example: 102
                  created_at:
                    type: integer
                    format: int64
                    description: "Record creation timestamp"
                    example: 1729614468
                  updated_at:
                    type: integer
                    format: int64
                    description: "Last updated timestamp"
                    example: 1729614468
                  owner_id:
                    type: integer
                    format: int64
                    description: "Foreign key, references the owners table"
                    example: 56
                  breed_id:
                    type: integer
                    format: int
                    description: "Foreign key, references the breeds table"
                    example: 10
                  gender_id:
                    type: string
                    description: "Foreign key to the genders table"
                    example: 2
                  coat_id:
                    type: integer
                    format: int
                    description: "Foreign key, references the coat table"
                    example: 7
                  name:
                    type: string
                    description: "Name of the shepherd dog"
                    example: "Scratch"
                  weight_g:
                    type: integer
                    format: int
                    description: "Record creation timestamp"
                    example: 30000
                  height_mm:
                    type: integer
                    format: int
                    description: "Record creation timestamp"
                    example: 10000
                  microchip_number:
                    type: integer
                    format: int
                    description: "microchip number"
                    example: 7
                  vaccinated:
                    type: boolean
                    description: "Indicates vaccination status"
                    example: true
                  date_of_birth:
                    type: integer
                    format: int64
                    description: "Dog’s date of birth"
                    example: 1697981263
        '400':
          $ref: "#/components/responses/InvalidRequest"
        '404':
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - dog
      summary: Create or update shepherd dog by ID
      description: Creates or updates a shepherd dog
      parameters:
        - name: id
          in: path
          description: ID of a shepherd dog to create or update
          required: true
          schema:
            type: integer
            format: int64
            example: 21
      requestBody:
        required: true
        description: Create or update an shepherd dog by ID
        content:
          application/json:
            schema:
              type: object
              properties:
                owner_id:
                  type: integer
                  format: int64
                  description: "Foreign key, references the owners table"
                  example: 56
                breed_id:
                  type: integer
                  format: int
                  description: "Foreign key, references the breeds table"
                  example: 10
                gender_id:
                  type: string
                  description: "Foreign key to the genders table"
                  example: 2
                coat_id:
                  type: integer
                  format: int
                  description: "Foreign key, references the coat table"
                  example: 7
                name:
                  type: string
                  description: "Name of the shepherd dog"
                  example: "Scratch"
                weight_g:
                  type: integer
                  format: int
                  description: "Record creation timestamp"
                  example: 30000
                height_mm:
                  type: integer
                  format: int
                  description: "Record creation timestamp"
                  example: 10000
                microchip_number:
                  type: integer
                  format: int
                  description: "microchip number"
                  example: 7
                vaccinated:
                  type: boolean
                  description: "Indicates vaccination status"
                  example: true
                date_of_birth:
                  type: integer
                  format: int64
                  description: "Dog’s date of birth"
                  example: 1697981263
      responses:
        '200':
          description: Successful operation
        '201':
          description: Successful creation
        '400':
          $ref: "#/components/responses/InvalidRequest"
    delete:
      tags:
        - dog
      summary: Delete shepherd dog by ID
      description: Deletes a shepherd dog
      parameters:
        - name: id
          in: path
          description: ID of a shepherd dog to delete
          required: true
          schema:
            type: integer
            format: int64
            example: 21
      responses:
        '204':
          description: Successful operation
        '400':
          $ref: "#/components/responses/InvalidRequest"
        '404':
          $ref: "#/components/responses/NotFound"
components:
  responses:
    InvalidRequest:
      description: The request was malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: "Error message"
          example: "config validation error: env 'HM_LOG_FORMAT' value 'textt' invalid, 'oneof=text json' expected;"